fix build without otlp-http or otlp-grpc
